Add HasAbstractCommitments trait, use it in PostRestartHtlcCleaner
